fix int2bytes, bytes2int
