+++ vm/neovm/types/integer.go
@@ -1,6 +1,7 @@
 package types
 
 import (
+	"github.com/dad-go/vm/neovm/interfaces"
 	"math/big"
 )
 
@@ -14,7 +15,7 @@ func NewInteger(value *big.Int) *Integer {
 	return &i
 }
 
-func (i *Integer) Equals(other StackItem) bool {
+func (i *Integer) Equals(other StackItemInterface) bool {
 	if _, ok := other.(*Integer); !ok {
 		return false
 	}
@@ -36,13 +37,13 @@ func (i *Integer) GetBoolean() bool {
 }
 
 func (i *Integer) GetByteArray() []byte {
-	return i.value.Bytes()
+	return ConvertBigIntegerToBytes(i.value)
 }
 
-func (i *Integer) GetInterface() {
-
+func (i *Integer) GetInterface() interfaces.IInteropInterface {
+	return nil
 }
 
-func (i *Integer) GetArray() []StackItem {
-	return []StackItem{i}
+func (i *Integer) GetArray() []StackItemInterface {
+	return []StackItemInterface{i}
 }
